+++ packages/@ionic/cli-utils/src/lib/events.ts
@@ -1,5 +1,0 @@
-import { EventEmitter } from 'events';
-
-import { ICLIEventEmitter } from '../definitions';
-
-export class CLIEventEmitter extends EventEmitter implements ICLIEventEmitter {}